Add an option for CCore to compile without multi threading support.
When the flag MULTI_THREADING is disabled, we run everything in one single thread.
We also make sure that we do not import any references to multi threading, so that
the Informatica Olympiade grading software can actually build and run the program.

Since the cache is faster for single threaded instances, we enable it by default.

Single threading does however mean that the progress counter does NOT work anymore.

Signed-off-by: Jord Nijhuis <jord@nijhuis.me>