Add header when generating CMerge content, Remove interactive mode, refactor
CMerge now inserts a small header containing the options that were passed to the program,
such as multi-threading and human mode.

The interactive mode was a bit interesting, since the interactive mode only would ask for
arguments that were not set. Because of this, we could never know if human mode was not set,
or if the user intentionally left it out. As such, we removed the interactive mode.

The file parameter has also been removed. Since we have no interactive mode, the only
printed content will be the source code, this means that you could just redirect it to a file
(e.g python3.5 src/CMerge c3 > c3.cpp)

Signed-off-by: Jord Nijhuis <jord@nijhuis.me>