Improved multi-threading support.
Previously, we divided the total amount of work by the amount of threads. This works fine,
if all the calculations take an equal amount of work. Unfortunately, this is not the case: The
higher the numbers are, the slower each calculation is. This means that the threads that handle
the higher numbers are still running when the other threads have already closed.

To fix this I implemented a WorkDivider, this class hands out batches to each thread, so that
- unless all batches are handed out- each thread has work to do. This increases our performance
for large ranges. For short ranges nothing really changes, since each thread will only receive one
batch.

Signed-off-by: Jord Nijhuis <jord@nijhuis.me>