Do a few optimizations regarding memory, and refactor some things.
Fixed that some pointers were not cleaned up afterwards. Also removed the list from the
progress counter, and instead replaced it by the last value we passed.

Also made sure that the work divider only accepts a tuple of <int, int> instead of two points,
as it only used the Y coordinate of those points. This also makes the divider more generic.

Signed-off-by: Jord Nijhuis <jord@nijhuis.me>